small changes to have a basic develop
I had to change a couple things around and remove some of the basic errors. I apologize for the @ts-ignore everywhere. I followed the older docs method and the new ones are more or less the same but with extra quirk but hopefully for the next project we'll use the up-to-date method. But you know: If it aint broke...